Abstracted pure functions and put them in a separate file
